Adding small client
I'm adding a small python client for the server-proxy. For the moment, we only have a stupid echo implementation.
The goal is to call the client from the sublime plugin, send a symbol, receive it at the proxy, and use the proxy to call the sbt server and request the information for this symbol.